requirements.txt
# Basics
numpy~=1.18.1
scipy==1.6.0
pandas==1.2.1
torch==1.7.1

networkx~=2.4
tqdm~=4.46.1
scikit-learn~=0.24.1
joblib~=0.15.1


# DGL
dgl-cu110==0.5.3

# Torch Geometric
torch-cluster==1.5.8
torch-geometric==1.6.3
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-spline-conv==1.2.0


# Visualization
seaborn==0.11.0
matplotlib==3.3.4
